requirements.txt
python-telegram-bot>=20.0,<22.0
python-dotenv>=1.0.0,<2.0.0
faster-whisper>=1.0.0,<2.0.0
speechbrain>=1.0.0
torchaudio>=2.0.0
torch>=2.0.0
numpy>=1.24.0
